adds explicit width and height to force layout on menu component
In some situations when a component that is passed to the menu prop and is wrapped
within a another Component view that provides no ability to style its view container
the menu view will collapse and hide its contents because the view container may not
have an explicit width and height set.

By adding an explicit width and height to the menu rule it will prevent this from
happening.