Don't queueAnimation if there is nothing to do
I'm not 100% sure why, but without this PR `<Navigator>` could have some issues with the default animation triggered by `pop()`. Somehow the animations interfere and the result is buggy/broken/ugly. 

A (somewhat hacky) solution for my case would be a delay infront of `Navigator.pop()`. But, IMHO, the solution in this PR is far superior, as we save some precious CPU cycles :)

Or is there a valid case to queue the open/close animation multiple times? I'm a bit confused, as there is already as state check in both `openMenu` and `closeMenu`, but only for the non-animation related stuff .. :|

(Tested with react-native 0.9.0-rc)