Make the open gesture claim the responder less greedily
The open gesture is currently very greedy: it sets itself the responder
whenever the touch moves more than 10px in the horizontal. This prevents other
components using horizontal swipe type interaction further down the component
hierarchy.

This commit changes this behaviour when the menu is closed. Responder is only
set if:
 * touch moves horizontally 10px (as before)
 * we're near the left edge of the screen (configurable)
 * we're swiping right