Fix glitch with initial render
This is a regression introduced in
https://github.com/react-native-fellowship/react-native-side-menu/commit/601cbe9d1333273fc829bf38475bee1f5487ab14
because of two things:

1. onLayout is called (according to documentation), immediately once the layout has been calculated,
but the new layout may not yet be reflected on the screen at the time the event is received,
especially if a layout animation is in progress.

2. State is missing initial width & height properties

We have managed to workaround that issue by adding setTimeout so that
the points 1 & 2 were temporarily hidden.

The purpose of onLayout was to recalculate width/height on orientation
change, not on initialRender as well.

Thanks to these changes, the glitch with initialRender no longer exists.

There is also another nice side-improvement -> when quickly refreshing -
the main content was glitching anyway - because there was few ms
difference between initialRender and onLayout being called - at the
first sight content view was 0x0 px in size - that was causing all the
elements to be in the top left corner. Now it no longer exists.