Allow configuring overlay styles in an animated fashion.
This PR supersedes #192 as it is strictly more expressive.

This PR allows the consumer to pass a function to style the overlay, just like the
one that styles the main wrapper view. The use-case for this is to allow e.g. fades
as you drag the menu over (it's a very nice effect!).

Because you can't style the component if it isn't mounted (i.e., during the initial
slide-out of the side menu), the overlay is now always rendered. It only captures
touch events if the menu is considered "open" so that the view behind the overlay
continues to work as normal while the menu is closed.

I also added animationStyle to propTypes and included an extra parameter to match the
new overlayAnimationStyle (the latter of which needs to know the maximum extent of
the pan gesture to animate a fade properly).